mediacapture-streams/MediaStreamTrack-getCapabilities.https.html

<!doctype html>
<title>MediaStreamTrack and InputDeviceInfo GetCapabilities</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script>

const audioProperties = [
  ["volume", "number"],
  ["sampleRate", "number"],
  ["sampleSize", "number"],
  ["echoCancellation", "boolean"],
  ["autoGainControl", "boolean"],
  ["noiseSuppression", "boolean"],
  ["latency", "number"],
  ["channelCount", "number"],
  ["deviceId", "string"],
  ["groupId", "string"]
];

const videoProperties = [
  ["width", "number"],
  ["height", "number"],
  ["aspectRatio", "number"],
  ["frameRate", "number"],
  ["facingMode", "enum", ["user", "environment", "left", "right"]],
  ["resizeMode", "enum", ["none", "crop-and-scale"]],
  ["deviceId", "string"],
  ["groupId", "string"]
];

function verifyBooleanCapability(capability) {
  assert_less_than_equal(capability.length, 2);
  capability.forEach(c => assert_equals(typeof c, "boolean"));
}

function verifyNumberCapability(capability) {
    assert_equals(typeof capability, "object");
    assert_equals(Object.keys(capability).length, 2);
    assert_true(capability.hasOwnProperty('min'));
    assert_true(capability.hasOwnProperty('max'));
    assert_less_than_equal(capability.min, capability.max);
}

function verifyEnumCapability(capability, enumMembers) {
  capability.forEach(c => {
    assert_equals(typeof c, "string");
    assert_in_array(c, enumMembers);
  });
}

function testCapabilities(capabilities, property, testNamePrefix) {
  const propertyName = property[0];
  const propertyType = property[1];
  let testName = testNamePrefix + " " + propertyName;
  test(() => {
    assert_true(capabilities.hasOwnProperty(propertyName));
  }, testName + " property present.");

  const capability = capabilities[propertyName];
  testName += " properly supported.";
  if (propertyType == "string") {
    test(() => {
      assert_equals(typeof capability, "string");
    }, testName);
  }

  if (propertyType == "boolean") {
    test(() => {
      verifyBooleanCapability(capability);
    }, testName);
  }

  if (propertyType == "number") {
    test(() => {
      verifyNumberCapability(capability);
    }, testName);
  }

  if (propertyType == "enum") {
    const valid_values = property[2];
    test(() => {
      verifyEnumCapability(capability, valid_values);
    }, testName);
  }
}

{
  audioProperties.forEach(property => {
    promise_test(async t => {
      const stream = await navigator.mediaDevices.getUserMedia({audio: true});
      t.add_cleanup(() => stream.getAudioTracks()[0].stop());
      const audioCapabilities = stream.getAudioTracks()[0].getCapabilities();
      testCapabilities(audioCapabilities, property, "Audio track getCapabilities()");
    }, "Setup audio MediaStreamTrack getCapabilities() test for " + property[0]);
  });

  videoProperties.forEach(property => {
    promise_test(async t => {
      const stream = await navigator.mediaDevices.getUserMedia({video: true});
      t.add_cleanup(() => stream.getVideoTracks()[0].stop());
      const audioCapabilities = stream.getVideoTracks()[0].getCapabilities();
      testCapabilities(audioCapabilities, property, "Video track getCapabilities()");
    }, "Setup video MediaStreamTrack getCapabilities() test for " + property[0]);
  });
}

{
  audioProperties.forEach(property => {
    promise_test(async t => {
      const devices = await navigator.mediaDevices.enumerateDevices();
      for (const device of devices) {
        // Test only one device.
        if (device.kind == "audioinput") {
          assert_inherits(device, "getCapabilities");
          const capabilities = device.getCapabilities();
          testCapabilities(capabilities, property, "Audio device getCapabilities()");
          break;
        }
      }
    }, "Setup audio InputDeviceInfo getCapabilities() test for " + property[0]);
  });

  videoProperties.forEach(property => {
    promise_test(async t => {
      const devices = await navigator.mediaDevices.enumerateDevices();
      for (const device of devices) {
        // Test only one device.
        if (device.kind == "videoinput") {
          assert_inherits(device, "getCapabilities");
          const capabilities = device.getCapabilities();
          testCapabilities(capabilities, property, "Video device getCapabilities()");
          break;
        }
      }
    }, "Setup video InputDeviceInfo getCapabilities() test for " + property[0]);
  });
}
</script>
